split form into seperate component
